Rename metricset to socket.
