+++ x-pack/auditbeat/include/list.go
@@ -10,6 +10,6 @@ import (
 	_ "github.com/elastic/beats/x-pack/auditbeat/module/system/host"
 	_ "github.com/elastic/beats/x-pack/auditbeat/module/system/packages"
 	_ "github.com/elastic/beats/x-pack/auditbeat/module/system/processes"
-	_ "github.com/elastic/beats/x-pack/auditbeat/module/system/sockets"
+	_ "github.com/elastic/beats/x-pack/auditbeat/module/system/socket"
 	_ "github.com/elastic/beats/x-pack/auditbeat/module/system/user"
 )
