+++ x-pack/auditbeat/module/system/socket/_meta/data.json
